Fix/2 separate service and controller logic

+++ src/main/java/com/interlink/auction/Controllers/AuctionController.java
@@ -1,12 +1,19 @@
 package com.interlink.auction.Controllers;
 
+import com.interlink.auction.Exceptions.InvalidBidValueException;
+import com.interlink.auction.Exceptions.WrongUserException;
 import com.interlink.auction.Models.DTO.AuctionDTORequest;
 import com.interlink.auction.Models.DTO.BidDTORequest;
 import com.interlink.auction.Models.Entities.Auction;
+import com.interlink.auction.Models.Entities.Bid;
+import com.interlink.auction.Models.Entities.Item;
 import com.interlink.auction.Services.AuctionService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:4200", allowedHeaders = "*", allowCredentials = "true")
@@ -23,33 +30,72 @@ public class AuctionController {
 
     @GetMapping("/{id}")
     public Auction getAuction(@PathVariable("id") Long id) {
-        return auctionService.getAuctionById(id);
+        try {
+            return auctionService.getAuctionById(id);
+        } catch (NullPointerException e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Auction not found with id: " + id);
+        }
     }
 
 
     @PostMapping()
     public Auction createAuction(@CookieValue(value = "id", defaultValue = "") String userId,
-                                 @RequestBody AuctionDTORequest auction) {
-        return auctionService.createAuction(userId, auction);
+                                 @RequestBody AuctionDTORequest auctionDTO) {
+        if (userId.equals(""))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User is not logged in");
+        try {
+            Item item = new Item();
+            item.setTitle(auctionDTO.getItemName());
+            Auction auction = new Auction(item, auctionDTO.getInitialBid(), LocalDateTime.now(), false);
+            return auctionService.createAuction(userId, auction);
+        } catch (NullPointerException e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found with id: " + userId);
+        }
     }
 
     @DeleteMapping("/{id}")
     public void deleteAuction(@CookieValue(value = "id", defaultValue = "") String userId,
                        @PathVariable("id") Long auctionId) {
-        auctionService.deleteAuction(userId, auctionId);
+        if (userId.equals("")) throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "User is not logged in");
+        try {
+            auctionService.deleteAuction(userId, auctionId);
+        } catch (NullPointerException e ) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Auction not found with id: " + userId);
+        } catch (WrongUserException e ) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, e.getMessage());
+        }
     }
 
     @PostMapping("/{id}/bids")
     public Auction makeBid(@CookieValue(value = "id", defaultValue = "") String userId,
                            @PathVariable("id") Long auctionId,
                            @RequestBody BidDTORequest bidDTORequest) {
-        return auctionService.makeBid(userId, auctionId, bidDTORequest);
+        if (userId.equals("")) throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "User is not logged in");
+        try {
+
+            Bid bid = new Bid(bidDTORequest.getBid(), bidDTORequest.getMaxBid(), null, null);
+            return auctionService.makeBid(userId, auctionId, bid);
+        } catch (NullPointerException e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
+        } catch (WrongUserException | InvalidBidValueException e) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, e.getMessage());
+        }
     }
 
     @PutMapping("/{id}")
     public Auction closeAuction(@CookieValue(value = "id", defaultValue = "") String userId,
                                 @PathVariable("id") Long auctionId,
                                 @RequestBody AuctionDTORequest auctionDTORequest) {
-        return auctionService.closeAuction(userId, auctionId, auctionDTORequest);
+        if (userId.equals("")) throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "User is not logged in");
+        try {
+            return auctionService.closeAuction(userId, auctionId, auctionDTORequest);
+        } catch (NullPointerException e ) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Auction not found with id: " + userId);
+        } catch (WrongUserException e ) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, e.getMessage());
+        }
     }
 }
